requirements.txt
tensorflow==2.14.0
opencv-python==4.8.1.78
paho-mqtt==1.6.1
numpy==1.26.1